Poly-commitment/Tests: move IPA related tests into ipa_commitment
Moving into tests subdirectory helps checking the API and give a context for
users to use the library.
