+++ poly-commitment/tests/ipa_commitment.rs
@@ -1,0 +1,333 @@
+use ark_ff::{One, UniformRand, Zero};
+use ark_poly::{
+    univariate::DensePolynomial, EvaluationDomain, Evaluations, Polynomial,
+    Radix2EvaluationDomain as D, Radix2EvaluationDomain, UVPolynomial,
+};
+use groupmap::GroupMap;
+use mina_curves::pasta::{Fp, Vesta as VestaG};
+use mina_poseidon::{
+    constants::PlonkSpongeConstantsKimchi as SC, sponge::DefaultFqSponge, FqSponge,
+};
+use o1_utils::ExtendedDensePolynomial;
+use poly_commitment::{
+    commitment::{
+        combine_evaluations, combined_inner_product, BatchEvaluationProof, CommitmentCurve,
+        Evaluation,
+    },
+    evaluation_proof::DensePolynomialOrEvaluations,
+    srs::SRS,
+    PolyComm, SRS as _,
+};
+use rand::{rngs::StdRng, SeedableRng};
+use std::array;
+
+#[test]
+fn test_combine_evaluations() {
+    let nb_of_chunks = 1;
+
+    // we ignore commitments
+    let dummy_commitments = PolyComm::<VestaG> {
+        elems: vec![VestaG::zero(); nb_of_chunks],
+    };
+
+    let polyscale = Fp::from(2);
+    // Using only one evaluation. Starting with eval_p1
+    {
+        let eval_p1 = Evaluation {
+            commitment: dummy_commitments.clone(),
+            evaluations: vec![
+                // Eval at first point. Only one chunk.
+                vec![Fp::from(1)],
+                // Eval at second point. Only one chunk.
+                vec![Fp::from(2)],
+            ],
+        };
+
+        let output = combine_evaluations::<VestaG>(&vec![eval_p1], polyscale);
+        // We have 2 evaluation points.
+        assert_eq!(output.len(), 2);
+        // polyscale is not used.
+        let exp_output = [Fp::from(1), Fp::from(2)];
+        output.iter().zip(exp_output.iter()).for_each(|(o, e)| {
+            assert_eq!(o, e);
+        });
+    }
+
+    // And after that eval_p2
+    {
+        let eval_p2 = Evaluation {
+            commitment: dummy_commitments.clone(),
+            evaluations: vec![
+                // Eval at first point. Only one chunk.
+                vec![Fp::from(3)],
+                // Eval at second point. Only one chunk.
+                vec![Fp::from(4)],
+            ],
+        };
+
+        let output = combine_evaluations::<VestaG>(&vec![eval_p2], polyscale);
+        // We have 2 evaluation points
+        assert_eq!(output.len(), 2);
+        // polyscale is not used.
+        let exp_output = [Fp::from(3), Fp::from(4)];
+        output.iter().zip(exp_output.iter()).for_each(|(o, e)| {
+            assert_eq!(o, e);
+        });
+    }
+
+    // Now with two evaluations
+    {
+        let eval_p1 = Evaluation {
+            commitment: dummy_commitments.clone(),
+            evaluations: vec![
+                // Eval at first point. Only one chunk.
+                vec![Fp::from(1)],
+                // Eval at second point. Only one chunk.
+                vec![Fp::from(2)],
+            ],
+        };
+
+        let eval_p2 = Evaluation {
+            commitment: dummy_commitments.clone(),
+            evaluations: vec![
+                // Eval at first point. Only one chunk.
+                vec![Fp::from(3)],
+                // Eval at second point. Only one chunk.
+                vec![Fp::from(4)],
+            ],
+        };
+
+        let output = combine_evaluations::<VestaG>(&vec![eval_p1, eval_p2], polyscale);
+        // We have 2 evaluation points
+        assert_eq!(output.len(), 2);
+        let exp_output = [Fp::from(1 + 3 * 2), Fp::from(2 + 4 * 2)];
+        output.iter().zip(exp_output.iter()).for_each(|(o, e)| {
+            assert_eq!(o, e);
+        });
+    }
+
+    // Now with two evaluations and two chunks
+    {
+        let eval_p1 = Evaluation {
+            commitment: dummy_commitments.clone(),
+            evaluations: vec![
+                // Eval at first point.
+                vec![Fp::from(1), Fp::from(3)],
+                // Eval at second point.
+                vec![Fp::from(2), Fp::from(4)],
+            ],
+        };
+
+        let eval_p2 = Evaluation {
+            commitment: dummy_commitments.clone(),
+            evaluations: vec![
+                // Eval at first point.
+                vec![Fp::from(5), Fp::from(7)],
+                // Eval at second point.
+                vec![Fp::from(6), Fp::from(8)],
+            ],
+        };
+
+        let output = combine_evaluations::<VestaG>(&vec![eval_p1, eval_p2], polyscale);
+        // We have 2 evaluation points
+        assert_eq!(output.len(), 2);
+        let o1 = Fp::from(1 + 3 * 2 + 5 * 4 + 7 * 8);
+        let o2 = Fp::from(2 + 4 * 2 + 6 * 4 + 8 * 8);
+        let exp_output = [o1, o2];
+        output.iter().zip(exp_output.iter()).for_each(|(o, e)| {
+            assert_eq!(o, e);
+        });
+    }
+}
+
+#[test]
+fn test_lagrange_commitments() {
+    let n = 64;
+    let domain = D::<Fp>::new(n).unwrap();
+
+    let mut srs = SRS::<VestaG>::create(n);
+    srs.add_lagrange_basis(domain);
+
+    let num_chunks = domain.size() / srs.g.len();
+
+    let expected_lagrange_commitments: Vec<_> = (0..n)
+        .map(|i| {
+            let mut e = vec![Fp::zero(); n];
+            e[i] = Fp::one();
+            let p = Evaluations::<Fp, D<Fp>>::from_vec_and_domain(e, domain).interpolate();
+            srs.commit_non_hiding(&p, num_chunks)
+        })
+        .collect();
+
+    let computed_lagrange_commitments = srs.lagrange_bases.get(&domain.size()).unwrap();
+    for i in 0..n {
+        assert_eq!(
+            computed_lagrange_commitments[i],
+            expected_lagrange_commitments[i],
+        );
+    }
+}
+
+#[test]
+// This tests with two chunks.
+fn test_chunked_lagrange_commitments() {
+    let n = 64;
+    let divisor = 4;
+    let domain = D::<Fp>::new(n).unwrap();
+
+    let mut srs = SRS::<VestaG>::create(n / divisor);
+    srs.add_lagrange_basis(domain);
+
+    let num_chunks = domain.size() / srs.g.len();
+    assert!(num_chunks == divisor);
+
+    let expected_lagrange_commitments: Vec<_> = (0..n)
+        .map(|i| {
+            let mut e = vec![Fp::zero(); n];
+            e[i] = Fp::one();
+            let p = Evaluations::<Fp, D<Fp>>::from_vec_and_domain(e, domain).interpolate();
+            srs.commit_non_hiding(&p, num_chunks)
+        })
+        .collect();
+
+    let computed_lagrange_commitments = srs.lagrange_bases.get(&domain.size()).unwrap();
+    for i in 0..n {
+        assert_eq!(
+            computed_lagrange_commitments[i],
+            expected_lagrange_commitments[i],
+        );
+    }
+}
+
+#[test]
+// TODO @volhovm I don't understand what this test does and
+// whether it is worth leaving.
+/// Same as test_chunked_lagrange_commitments, but with a slight
+/// offset in the SRS
+fn test_offset_chunked_lagrange_commitments() {
+    let n = 64;
+    let domain = D::<Fp>::new(n).unwrap();
+
+    let mut srs = SRS::<VestaG>::create(n / 2 + 1);
+    srs.add_lagrange_basis(domain);
+
+    // Is this even taken into account?...
+    let num_chunks = (domain.size() + srs.g.len() - 1) / srs.g.len();
+    assert!(num_chunks == 2);
+
+    let expected_lagrange_commitments: Vec<_> = (0..n)
+        .map(|i| {
+            let mut e = vec![Fp::zero(); n];
+            e[i] = Fp::one();
+            let p = Evaluations::<Fp, D<Fp>>::from_vec_and_domain(e, domain).interpolate();
+            srs.commit_non_hiding(&p, num_chunks) // this requires max = Some(64)
+        })
+        .collect();
+
+    let computed_lagrange_commitments = srs.lagrange_bases.get(&domain.size()).unwrap();
+    for i in 0..n {
+        assert_eq!(
+            computed_lagrange_commitments[i],
+            expected_lagrange_commitments[i],
+        );
+    }
+}
+
+#[test]
+fn test_opening_proof() {
+    // create two polynomials
+    let coeffs: [Fp; 10] = array::from_fn(|i| Fp::from(i as u32));
+    let poly1 = DensePolynomial::<Fp>::from_coefficients_slice(&coeffs);
+    let poly2 = DensePolynomial::<Fp>::from_coefficients_slice(&coeffs[..5]);
+
+    // create an SRS
+    let srs = SRS::<VestaG>::create(20);
+    let rng = &mut StdRng::from_seed([0u8; 32]);
+
+    // commit the two polynomials
+    let commitment1 = srs.commit(&poly1, 1, rng);
+    let commitment2 = srs.commit(&poly2, 1, rng);
+
+    // create an aggregated opening proof
+    let (u, v) = (Fp::rand(rng), Fp::rand(rng));
+    let group_map = <VestaG as CommitmentCurve>::Map::setup();
+    let sponge = DefaultFqSponge::<_, SC>::new(mina_poseidon::pasta::fq_kimchi::static_params());
+
+    let polys: Vec<(
+        DensePolynomialOrEvaluations<_, Radix2EvaluationDomain<_>>,
+        PolyComm<_>,
+    )> = vec![
+        (
+            DensePolynomialOrEvaluations::DensePolynomial(&poly1),
+            commitment1.blinders,
+        ),
+        (
+            DensePolynomialOrEvaluations::DensePolynomial(&poly2),
+            commitment2.blinders,
+        ),
+    ];
+    let elm = vec![Fp::rand(rng), Fp::rand(rng)];
+
+    let opening_proof = srs.open(&group_map, &polys, &elm, v, u, sponge.clone(), rng);
+
+    // evaluate the polynomials at these two points
+    let poly1_chunked_evals = vec![
+        poly1
+            .to_chunked_polynomial(1, srs.g.len())
+            .evaluate_chunks(elm[0]),
+        poly1
+            .to_chunked_polynomial(1, srs.g.len())
+            .evaluate_chunks(elm[1]),
+    ];
+
+    fn sum(c: &[Fp]) -> Fp {
+        c.iter().fold(Fp::zero(), |a, &b| a + b)
+    }
+
+    assert_eq!(sum(&poly1_chunked_evals[0]), poly1.evaluate(&elm[0]));
+    assert_eq!(sum(&poly1_chunked_evals[1]), poly1.evaluate(&elm[1]));
+
+    let poly2_chunked_evals = vec![
+        poly2
+            .to_chunked_polynomial(1, srs.g.len())
+            .evaluate_chunks(elm[0]),
+        poly2
+            .to_chunked_polynomial(1, srs.g.len())
+            .evaluate_chunks(elm[1]),
+    ];
+
+    assert_eq!(sum(&poly2_chunked_evals[0]), poly2.evaluate(&elm[0]));
+    assert_eq!(sum(&poly2_chunked_evals[1]), poly2.evaluate(&elm[1]));
+
+    let evaluations = vec![
+        Evaluation {
+            commitment: commitment1.commitment,
+            evaluations: poly1_chunked_evals,
+        },
+        Evaluation {
+            commitment: commitment2.commitment,
+            evaluations: poly2_chunked_evals,
+        },
+    ];
+
+    let combined_inner_product = {
+        let es: Vec<_> = evaluations
+            .iter()
+            .map(|Evaluation { evaluations, .. }| evaluations.clone())
+            .collect();
+        combined_inner_product(&v, &u, &es)
+    };
+
+    // verify the proof
+    let mut batch = vec![BatchEvaluationProof {
+        sponge,
+        evaluation_points: elm.clone(),
+        polyscale: v,
+        evalscale: u,
+        evaluations,
+        opening: &opening_proof,
+        combined_inner_product,
+    }];
+
+    assert!(srs.verify(&group_map, &mut batch, rng));
+}
